implemented and fixed retargeting feature of followbullet

+++ TowerDefenseGame/Projectiles.pde
@@ -1,13 +1,13 @@
 abstract class Projectiles {
-  float vx,vy, x,y, damage,speed;
+  float vx, vy, x, y, damage, speed;
   int penetrationLevel, size;
   boolean canAttackArmored, doneShooting;
 
 
-  Projectiles(float xA, float yA, Monster i, float damageA){
+  Projectiles(float xA, float yA, Monster i, float damageA) {
     speed = 4;
-    vx = ((i.x - xA)/ (float)Math.sqrt(Math.pow(i.x - xA,2) + Math.pow(i.y - yA,2))) * speed;
-    vy = ((i.y - yA)/ (float)Math.sqrt(Math.pow(i.x - xA,2) + Math.pow(i.y - yA,2))) * speed;
+    vx = ((i.x - xA)/ (float)Math.sqrt(Math.pow(i.x - xA, 2) + Math.pow(i.y - yA, 2))) * speed;
+    vy = ((i.y - yA)/ (float)Math.sqrt(Math.pow(i.x - xA, 2) + Math.pow(i.y - yA, 2))) * speed;
     size = 20;
     canAttackArmored = false;
     x = xA;
@@ -17,60 +17,80 @@ abstract class Projectiles {
     damage = damageA;
     //level = 1;
   }
-  boolean dealDamage(Monster i){
-    if(Math.pow(i.x - x,2) + Math.pow(i.y - y,2) <= Math.pow(size,2)){ //monster is in bullet's range
+  boolean dealDamage(Monster i) {
+    if (Math.pow(i.x - x, 2) + Math.pow(i.y - y, 2) <= Math.pow(size, 2)) { //monster is in bullet's range
       i.changeHP(-1 * damage);
       penetrationLevel--;
-      if(penetrationLevel <= 0){ //if pentration level dips below 0, kill the bullet
+      if (penetrationLevel <= 0) { //if pentration level dips below 0, kill the bullet
         toDestroyA.add(this);
       }
       return true;
     }
     return false;
   }
-  void move(){
+  void move() {
     x += vx;
     y += vy;
-    for(Monster m: Monsters){
-        if(this.x < 0 || this.x > 1280 || this.y < 0 || this.y > 720){
+    for (Monster m : Monsters) {
+      if (this.x < 0 || this.x > 1280 || this.y < 0 || this.y > 720) {
         toDestroyA.add(this);
         break;
-        }
-        if(this.dealDamage(m)){
+      }
+      if (this.dealDamage(m)) {
         break;
-        }
+      }
     }
   }
 }
 
-class StraightBullet extends Projectiles{
-  StraightBullet(float xA, float yA, Monster i, float damage){
+class StraightBullet extends Projectiles {
+  StraightBullet(float xA, float yA, Monster i, float damage) {
     super(xA, yA, i, damage);
   }
-  void move(){
+  void move() {
     super.move();
   }
 }
 
-class followBullet extends Projectiles{
+class followBullet extends Projectiles {
   float turnedTime;
   Monster monster;
   boolean resting;
-  followBullet(float xA, float yA, Monster i, float damage){
+  followBullet(float xA, float yA, Monster i, float damage) {
     super(xA, yA, i, damage);
     monster = i;
     resting = true;
     turnedTime =  millis();
   }
-  void move(){
+  void move() {
+    if (monster == null) {
+      System.out.println("heyyy");
+      monster = Monsters.get(0);
+    }
     super.move(); //move foward
-    if(resting && (millis() - turnedTime)/70 >= 1){ //if it's resting and .07 seconds passed since it was redirected, say it's not resting
-        resting = false;
+    if (resting && (millis() - turnedTime)/70 >= 1) { //if it's resting and .07 seconds passed since it was redirected, say it's not resting
+      resting = false;
     }
-    if(!resting){ //if it's not resting, redirect the bullet towards the monster and then set the time it was turned
-        vx = ((monster.x - x)/ (float)Math.sqrt(Math.pow(monster.x - x,2) + Math.pow(monster.y - y,2))) * speed;
-        vy = ((monster.y - y)/ (float)Math.sqrt(Math.pow(monster.x - x,2) + Math.pow(monster.y - y,2))) * speed;
-        turnedTime = millis();
+    if (!resting) { //if it's not resting, redirect the bullet towards the monster and then set the time it was turned
+      vx = ((monster.x - x)/ (float)Math.sqrt(Math.pow(monster.x - x, 2) + Math.pow(monster.y - y, 2))) * speed;
+      vy = ((monster.y - y)/ (float)Math.sqrt(Math.pow(monster.x - x, 2) + Math.pow(monster.y - y, 2))) * speed;
+      turnedTime = millis();
     }
+  }
+
+  boolean dealDamage(Monster i) {
+    if (Math.pow(i.x - x, 2) + Math.pow(i.y - y, 2) <= Math.pow(size, 2)) { //monster is in bullet's range
+      if (i.changeHP(-1 * damage) <=0) {
+        monster = null;
+        penetrationLevel--;
+        if (penetrationLevel <= 0) { //if pentration level dips below 0, kill the bullet
+          toDestroyA.add(this);
+        }
+      } else{ //if bullet wasn't able to the monster completely, kill the bullet
+        toDestroyA.add(this);
+      }
+      return true;
+    }
+    return false;
   }
 }
